refactor: Group all `src/**/*.ts` into subfolders Groups are: `web/`, `node/`, `common/`
