+++ src/common/cheatsheet/cheatsheet-content.ts
