+++ src/node/export/variable-resolver.ts
